+++ frontend/src/main.ts
@@ -38,7 +38,6 @@ export const tags = fetch("/api/tags").then(async (response: any) =>
 	})
 );
 const clientId = import.meta.env.VITE_DISCORD_ID ?? "";
-console.log(clientId);
 const loginLink = `https://discord.com/api/oauth2/authorize?client_id=${clientId}&response_type=code&scope=identify+email&redirect_uri=${encodeURIComponent(window.location.origin + "/user")}`;
 
 export function sendToLogin(route: string) {